requirements.txt
altgraph==0.17.4
attrs==23.2.0
auto-py-to-exe==2.35.0
beautifulsoup4==4.12.3
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.2
camelot-py==0.11.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.8
Eel==0.16.0
et-xmlfile==1.1.0
future==1.0.0
get-chrome-driver==1.3.20
gevent==24.2.1
gevent-websocket==0.10.1
ghostscript==0.7
greenlet==3.0.3
h11==0.14.0
humanfriendly==10.0
idna==3.7
img2table==1.2.11
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
#numpy==1.24.4
opencv-python==4.7.0.72
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdf2image==1.16.3
pdfminer.six==20240706
pefile==2023.2.7
pillow==10.4.0
polars==1.1.0
pyarrow==16.1.0
pycparser==2.22
Pygments==2.18.0
#pyinstaller==5.11.0
pyinstaller-hooks-contrib==2024.7
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==2.12.1
pypyodbc==1.3.6
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
redmail==0.6.0
requests==2.32.3
rich==13.7.1
selenium==4.22.0
selenium-stealth==1.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabulate==0.9.0
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==3.8.6
websocket-client==1.8.0
whichcraft==0.6.1
wsproto==1.2.0
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.4.post2
